markdownToRich: hack tree problem by hooking directly into the lexer
We encountered an issue where, since we were immediately appending to
the MessageEmbed object as the renderer was passed tokens, we were
displaying duplicate data when, e.g. a 'blockquote' element's children
were interpreted as paragraphs.

The solution was to hook directly into the lexer, using the AST they
build to construct a message from only the parts we need. This is VERY
brittle and has not been tested with all inline and block element types.
If no strategy to handle an element is present, the raw text is appended
to the message directly.